Split the sc trade tools client
